+++ tests/testthat/resources/parsnip.R
@@ -1,0 +1,3 @@
+mod <-
+    logistic_reg(mixture = 1/3) %>%
+    set_engine("glmnet", nlambda = 10)